refactor: 구매 로또와 당첨 결과 번호를 분리
