+++ src/main/java/lotto/controller/GameController.java
@@ -1,9 +1,6 @@
 package lotto.controller;
 
-import lotto.domain.LottoPaper;
-import lotto.domain.LottoNumberStragey;
-import lotto.domain.RandomNumber;
-import lotto.domain.Statistics;
+import lotto.domain.*;
 import lotto.view.InputView;
 import lotto.view.ResultView;
 
@@ -13,12 +10,12 @@ import java.util.List;
 public class GameController {
 
     private List<LottoPaper> buyLotto;
-    private LottoPaper winningNumber;
+    private WinningLottoNumber winningNumber;
     private Statistics statistics;
 
     public GameController() {
         buyLotto = new ArrayList<>();
-        winningNumber = new LottoPaper();
+        winningNumber = new WinningLottoNumber();
         statistics = new Statistics();
     }
 
@@ -35,17 +32,9 @@ public class GameController {
 
         resultView.lottoNumberView(buyLotto);
         winningNumber.setWinningNumber(inputView.winningLottoView());
-        checkLottoStatitics(buyLottoMoney);
+
         resultView.winningResult(statistics);
     }
 
-    private void checkLottoStatitics(int buyMoney) {
-
-        for (LottoPaper lotto : buyLotto) {
-            lotto.compreWinningNumber(winningNumber.getLottoNumber(), statistics);
-        }
-
-        statistics.setEarnMoneyPercentage(buyMoney);
-    }
 
 }
